Improvements to migration logic

+++ beta/MXWeather.sh
@@ -4,27 +4,32 @@
 
 # Enables migration if the environment variable is set
 if [ "$MIGRATE" != "false" ]; then
-echo "Migration enabled. Checking if migration has already been completed..."
+echo "Migration enabled. Begin migration checks..."
 
   # Checks if there is more than 1 file in the data folder (indicates new install or existing)
   if [ "$(ls -A /opt/CumulusMX/data/ | wc -l)" -gt 1 ]; then 
-  echo "Multiple files detected. Proceeding with migration..."
+  echo "Multiple files detected. Checking if migration has already been completed..."
 
-    # Checks to see if data has already been migrated and skips if it has. 
+    # Checks to see if data has already been migrated and skips migration if it has. 
     if [ ! -f "/opt/CumulusMX/config/.migrated" ] && [ ! -f "/opt/CumulusMX/config/.nodata" ] || [ "$MIGRATE" == "force" ]; then 
-      echo "Migration has not been completed. Starting migration..."
+      if [ "$MIGRATE" == "force" ]; then
+        echo "Migration is being forced. Backing up files..."
+      else 
+       echo "No previous migration detected. Backing up files..."
+      fi
 
       # Backup Cumulus.ini file
       cp -R /opt/CumulusMX/config/Cumulus.ini /opt/CumulusMX/config/Cumulus-v3.ini.bak
       echo "Backed up Cumulus.ini"
 
-      # Copy Cumulus.ini to root
-      cp -f /opt/CumulusMX/config/Cumulus.ini /opt/CumulusMX/
-
       # Backup data files
       mkdir -p /opt/CumulusMX/backup/datav3
       cp -R /opt/CumulusMX/data/* /opt/CumulusMX/backup/datav3
       echo "Backed up data folder"
+
+      # Copy Cumulus.ini to root
+      cp -f /opt/CumulusMX/config/Cumulus.ini /opt/CumulusMX/
+      echo "Copied Cumulus.ini file to root"
 
       # Copy data files to datav3 for migration
       mkdir /opt/CumulusMX/datav3
@@ -32,6 +37,7 @@ echo "Migration enabled. Checking if migration has already been completed..."
       echo "Copied data files to migration folder"
 
       # Run migration script
+      echo "Running migration task..."
       expect <<EOF
 spawn dotnet MigrateData3to4.dll
 expect "Press a Enter to continue, or Ctrl-C to exit"
@@ -44,13 +50,16 @@ EOF
       # Leave a file to indicate the migration has been completed
       touch /opt/CumulusMX/config/.migrated
 
-      # Copy UniqueID file to config folder
-      cp -f /opt/CumulusMX/UniqueId.txt /opt/CumulusMX/config/
+      # Copy UniqueID file to config folder if it exists
+
+      if [ -f "/opt/CumulusMX/UniqueId.txt" ]; then
+        cp -f /opt/CumulusMX/UniqueId.txt /opt/CumulusMX/config/
+      fi
       echo "Migration completed. Starting CumulusMX..."
 
     else 
       # If the .migrated or .nodata file already exists it skips the migration. 
-      echo "Migration already completed... Skipping migration."
+      echo "Migration has already been done... Skipping migration."
     fi
 
   else
